Item10803: Fix broken Config.spec.
Sadly, I had to add a configure checker for all the existing installations
we've shipped the broken Config.spec to. It will flag an error and provide
instructions to replace the illegal value by copy-pasting from the previous
switchboard entry.

The problem was that I forgot to add any type info for the compareauth
switchboard entry, so Foswiki configure assumed it was a string instead of a
hash.

Although this is a terrible mistake on my part, this switchboard entry is only used when:
   * You are using ApacheLogin
   * You are not authenticated
   * You directly access an /bin/compare or /bin/compareauth URL

Oops :-(

git-svn-id: http://svn.foswiki.org/trunk/CompareRevisionsAddOn@11783 0b4bb1d4-4e5a-0410-9cc4-b2b747904278