Item11755: fix by completely disabling all entity processing
If we don't let HTML::TreeBuilder decode entities in the first place, we know
that all text nodes contain proper HTML entities and so we can skip
re-encoding things. This means we remove potentially fault-inducing steps of
processing which is a Good Thing™ and certainly beats the previous attempt at
fixing the problem in every way.

git-svn-id: http://svn.foswiki.org/branches/Release01x01/CompareRevisionsAddOn@15832 0b4bb1d4-4e5a-0410-9cc4-b2b747904278